Excluded the specific helper methods into separate classes.
